generate-password/requirements.txt
psycopg2-binary
bcrypt
qrcode
pillow